reorg multiclass
